redo hook for using local samples
