+++ src/components/Sampler/style.sass
@@ -11,7 +11,5 @@ $sampler-height: calc(100vh - #{$header-height})
 
 
 #controller
-  // height: rem(600)
-  // max-height: $sampler-height
   +phone-portrait
     height: $sampler-height
