+++ src/index.tsx
@@ -1,5 +1,5 @@
 // Must be the first import
-// if (process.env.NODE_ENV === 'development') {
+// if (devOnly) {
 // Must use require here as import statements are only allowed to exist at top-level.
 // require('preact/debug');
 // }
