+++ src/utils.sass
@@ -6,6 +6,9 @@ $browser-context: 16 // Chrome default font size in px
 @function rem($size)
   @return ($size / $browser-context) * 1rem
 
+/*************************************
+  CSS variables
+  ***********************************/
 :root
   --pad-min-size: 140px
