+++ src/utils.ts
@@ -1,0 +1,1 @@
+export const devOnly = process.env.NODE_ENV === 'development';
